refactor: move getProfileHeaderOption to utils
