+++ common/utils/getProfileHeaderOption.js
